requirements.txt
httpx>=0.27.0
supabase>=2.4.0
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
websockets>=12.0
rich>=13.7.0
